backend/app/requirements.txt
fastapi
pydantic
pydantic[email]
pydantic[dotenv]
uvicorn
neo4j
redis
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
pytest-cov